// packages/pivo/src/open-api/open-api-types.ts
  interface ParameterObject extends ParameterBaseObject {
    name: string
    in: string
  }
  interface HeaderObject extends ParameterBaseObject {}
  type ParameterBaseObject = {
    description?: string
    required?: boolean
    deprecated?: boolean
    allowEmptyValue?: boolean
    style?: string
    explode?: boolean
    allowReserved?: boolean
    schema?: ReferenceObject | SchemaObject
    example?: any
    examples?: {
      [media: string]: ReferenceObject | ExampleObject
    }
    content?: {
      [media: string]: MediaTypeObject
    }
  } & JsonLD.Entry
  type NonArraySchemaObjectType =
    | 'null'
    | 'boolean'
    | 'object'
    | 'number'
    | 'string'
    | 'integer'
  type ArraySchemaObjectType = 'array'
  type SchemaObject = ArraySchemaObject | NonArraySchemaObject
  interface ArraySchemaObject extends BaseSchemaObject {
    type: ArraySchemaObjectType
    items: ReferenceObject | SchemaObject
  }
  interface NonArraySchemaObject extends BaseSchemaObject {
    type: NonArraySchemaObjectType
  }
// ...
  interface ContentsObject {
    [media: string]: MediaTypeObject
  }
  interface RequestBodyObject {
    description?: string
    content: ContentsObject
    required?: boolean
  }
// ...
export declare namespace ExpandedOpenAPIV3Semantics {
// ...
  interface PathItemObject {
    $ref?: string
    summary?: string
    description?: string
    get?: OperationObject
    put?: OperationObject
    post?: OperationObject
    delete?: OperationObject
    options?: OperationObject
    head?: OperationObject
    patch?: OperationObject
    trace?: OperationObject
    servers?: ServerObject[]
    parameters?: Array<ParameterObject>
  }
  type OperationObject = {
    url: string
    verb: string
    tags?: string[]
    summary?: string
    description?: string
    externalDocs?: ExternalDocumentationObject
    operationId?: string
    parameters?: Array<ParameterObject>
    requestBody?: RequestBodyObject
    responses?: ResponsesObject
    callbacks?: {
      [callback: string]: CallbackObject
    }
    deprecated?: boolean
    security?: SecurityRequirementObject[]
    servers?: ServerObject[]
    userShouldAuthenticate?: boolean
  } & JsonLD.Entry
// ...
  interface ArraySchemaObject extends BaseSchemaObject {
    type: ArraySchemaObjectType
    items: SchemaObject
  }
  interface NonArraySchemaObject extends BaseSchemaObject {
    type: NonArraySchemaObjectType
  }
  type BaseSchemaObject = {
// ...
    'x-affiliation'?: 'parent' | 'this' // addition
  } & JsonLD.Entry
  interface DiscriminatorObject {
    propertyName: string
    mapping?: {
      [value: string]: string
    }
  }
  interface XMLObject {
    name?: string
    namespace?: string
    prefix?: string
    attribute?: boolean
    wrapped?: boolean
  }
  interface ExampleObject {
    summary?: string
    description?: string
    value?: any
    externalValue?: string
  }
// ...
}